backend/requirements1.txt
langchain
bs4
fastapi
requests
uvicorn
python-dotenv
opencv-python
pymongo
tensorflow
python-multipart